# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fuzzbin API",
    "description": "Music video library management API.\nProvides CRUD operations for videos, artists, collections, and tags,\nwith full-text search and filtering capabilities.\n\n## Authentication\n\nWhen authentication is enabled (`FUZZBIN_API_AUTH_ENABLED=true`), protected endpoints require\na Bearer token in the `Authorization` header:\n\n```\nAuthorization: Bearer <your-jwt-token>\n```\n\nObtain a token via `POST /auth/token` with username/password credentials.\n\n## WebSocket API\n\n### Real-time Job Progress: `/ws/jobs/{job_id}`\n\nConnect to receive live progress updates for background jobs.\n\n**Connection:**\n```\nws://localhost:8000/ws/jobs/{job_id}\n```\n\n**Message Format (JSON):**\n```json\n{\n  \"job_id\": \"uuid-string\",\n  \"status\": \"running\",\n  \"progress\": 0.45,\n  \"message\": \"Processing file 45 of 100\",\n  \"result\": null\n}\n```\n\n**Fields:**\n- `job_id`: Job identifier\n- `status`: One of `pending`, `running`, `completed`, `failed`, `cancelled`\n- `progress`: Float 0.0-1.0 indicating completion percentage\n- `message`: Human-readable status message\n- `result`: Job result data (populated on completion)\n\n**Close Codes:**\n- `1008` (Policy Violation): Invalid or unknown job ID\n- `1011` (Internal Error): Server error during progress streaming\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/jobs/{job_id}": {
      "get": {
        "tags": [
          "Jobs"
        ],
        "summary": "Get job status",
        "description": "Get the current status and progress of a job.",
        "operationId": "get_job_jobs__job_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Jobs"
        ],
        "summary": "Cancel a job",
        "description": "Cancel a pending or running job. Has no effect on completed jobs.",
        "operationId": "cancel_job_jobs__job_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/backup": {
      "get": {
        "tags": [
          "Backup"
        ],
        "summary": "List backups",
        "description": "List all available backup archives.\n\nReturns backup metadata including filename, size, creation timestamp,\nand contents. Results are sorted by creation time (newest first).",
        "operationId": "list_backups_backup_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions or account disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Backup"
        ],
        "summary": "Create backup",
        "description": "Trigger an on-demand system backup.\n\nCreates a background job that generates a .zip archive containing:\n- **config.yaml**: User configuration file\n- **fuzzbin.db**: Library database (using SQLite backup API)\n- **.thumbnails/**: Cached thumbnail images\n\nThe backup job runs asynchronously. Use the returned `job_id` to track\nprogress via `GET /jobs/{job_id}` or WebSocket `/ws/jobs/{job_id}`.\n\nOld backups exceeding the retention count are automatically deleted.",
        "operationId": "create_backup_backup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BackupCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupCreateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions or account disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error - Unexpected server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/backup/{filename}": {
      "get": {
        "tags": [
          "Backup"
        ],
        "summary": "Download backup",
        "description": "Download a backup archive by filename.\n\nReturns the .zip file as a binary download. The archive can be\nextracted and restored manually without the program running.",
        "operationId": "download_backup_backup__filename__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "401": {
            "description": "Unauthorized - Authentication required or token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions or account disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/backup/{filename}/verify": {
      "get": {
        "tags": [
          "Backup"
        ],
        "summary": "Verify backup integrity",
        "description": "Verify a backup archive's integrity.\n\nChecks:\n- ZIP file structure and CRC checksums\n- Database SQLite integrity (if database is included)\n- Presence of expected files\n\nReturns verification results including any errors found.",
        "operationId": "verify_backup_backup__filename__verify_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Filename"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BackupVerifyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized - Authentication required or token invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Insufficient permissions or account disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "404": {
            "description": "Not Found - Resource does not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BackupCreateRequest": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional description for the backup",
            "examples": [
              "Pre-upgrade backup"
            ]
          }
        },
        "type": "object",
        "title": "BackupCreateRequest",
        "description": "Request to create a new backup."
      },
      "BackupCreateResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id",
            "description": "ID of the background backup job",
            "examples": [
              "550e8400-e29b-41d4-a716-446655440000"
            ]
          },
          "message": {
            "type": "string",
            "title": "Message",
            "description": "Human-readable status message",
            "examples": [
              "Backup job submitted successfully"
            ]
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "message"
        ],
        "title": "BackupCreateResponse",
        "description": "Response when a backup job is submitted.\n\nThe backup runs as a background job. Poll the job status endpoint\nfor progress and completion."
      },
      "BackupInfo": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Name of the backup file (e.g., 'fuzzbin_backup_20251222_140530.zip')",
            "examples": [
              "fuzzbin_backup_20251222_140530.zip"
            ]
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes",
            "description": "Size of the backup archive in bytes",
            "examples": [
              15728640
            ]
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "description": "ISO 8601 timestamp when the backup was created",
            "examples": [
              "2025-12-22T14:05:30+00:00"
            ]
          },
          "contains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contains",
            "description": "List of items included in backup: 'config', 'database', 'thumbnails'",
            "examples": [
              [
                "config",
                "database",
                "thumbnails"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "filename",
          "size_bytes",
          "created_at",
          "contains"
        ],
        "title": "BackupInfo",
        "description": "Information about a backup archive.\n\nReturned when listing or creating backups."
      },
      "BackupListResponse": {
        "properties": {
          "backups": {
            "items": {
              "$ref": "#/components/schemas/BackupInfo"
            },
            "type": "array",
            "title": "Backups",
            "description": "List of available backup archives, sorted by creation time (newest first)"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total number of backups available",
            "examples": [
              7
            ]
          }
        },
        "type": "object",
        "required": [
          "backups",
          "total"
        ],
        "title": "BackupListResponse",
        "description": "Response containing list of available backups."
      },
      "BackupVerifyResponse": {
        "properties": {
          "valid": {
            "type": "boolean",
            "title": "Valid",
            "description": "True if backup passed all integrity checks"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Name of the verified backup file"
          },
          "contains": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Contains",
            "description": "List of items found in backup: 'config', 'database', 'thumbnails'"
          },
          "database_valid": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Database Valid",
            "description": "True if database passed SQLite integrity check (null if no database in backup)"
          },
          "errors": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Errors",
            "description": "List of error messages if verification failed"
          }
        },
        "type": "object",
        "required": [
          "valid",
          "filename",
          "contains"
        ],
        "title": "BackupVerifyResponse",
        "description": "Response from backup verification."
      },
...
      "ErrorDetail": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "description": "Error message"
          },
          "error_type": {
            "type": "string",
            "title": "Error Type",
            "description": "Error type identifier"
          }
        },
        "type": "object",
        "required": [
          "detail",
          "error_type"
        ],
        "title": "ErrorDetail",
        "description": "Standard error response detail."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type": {
            "$ref": "#/components/schemas/JobType"
          },
          "status": {
            "$ref": "#/components/schemas/JobStatus"
          },
          "progress": {
            "type": "number",
            "title": "Progress"
          },
          "current_step": {
            "type": "string",
            "title": "Current Step"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "processed_items": {
            "type": "integer",
            "title": "Processed Items"
          },
          "result": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "metadata": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type",
          "status",
          "progress",
          "current_step",
          "total_items",
          "processed_items",
          "result",
          "error",
          "created_at",
          "started_at",
          "completed_at",
          "metadata"
        ],
        "title": "JobResponse",
        "description": "Job status response.\n\nContains all job information including progress and results."
      },
      "JobStatus": {
        "type": "string",
        "enum": [
          "pending",
          "waiting",
          "running",
          "completed",
          "failed",
          "cancelled",
          "timeout"
        ],
        "title": "JobStatus",
        "description": "Job status enumeration."
      },
...
      "JobType": {
        "type": "string",
        "enum": [
          "import_nfo",
          "import_spotify",
          "download_youtube",
          "file_organize",
          "file_duplicate_resolve",
          "metadata_enrich",
          "metadata_refresh",
          "library_scan",
          "import",
          "backup"
        ],
        "title": "JobType",
        "description": "Job type enumeration."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from POST /auth/token"
      }
    }
  },
...
}